+++ ChangeLog.md
@@ -19,6 +19,7 @@
 - Refactoring: Modul umbenannt serde.rs -> speichern.rs
 - Refactoring: Anpassung diverser Traits an neue Anforderungen
 - Refactoring: GleisId verwendet u32 statt u64 (raspi ist 32Bit System)
+- Refactoring: Modul umbenannt polarity.rs -> polarität.rs
 
 ## 2.0.1